Moved some EquiZ detail into separate file.
